cicd: keep the changelog header version in step with the release tag
- The changelog writer filled the release header from a SemVer bump of the last version (a feature is a minor bump), which disagreed with the date-based version the release actually tags — so a release that contained a feature previewed the wrong number in the changelog header (the file it wrote was already correct; only the preview was off). A shared finalizeContext now rewrites the header to the same date-based version the tag uses. The date-version calculation moved into one shared helper so the version plugin and the changelog can never drift apart.
